feat: Rename project from SerialLCD to LifelineTTY
- Updated all references from SerialLCD to LifelineTTY across the codebase, including binaries, service files, and documentation.
- Added compatibility symlink for legacy support, allowing existing scripts to continue functioning with the old name.
- Adjusted packaging scripts to reflect the new binary name and service unit.
- Enhanced README and migration notes to guide users through the transition.
- Maintained CLI behavior and configuration file paths for backward compatibility.

+++ src/display/overlays.rs
@@ -134,7 +134,7 @@ fn render_bar(percent: u8, width: usize) -> String {
         return String::new();
     }
 
-    let max_level = (BAR_LEVELS.len() - 1) as usize;
+    let max_level = BAR_LEVELS.len() - 1;
     let total_units = width * max_level;
     let filled_units = (percent as usize * total_units) / 100;
     let mut s = String::with_capacity(width);
